Port remaining test, visibility cleanup
